requirements.txt
pygwan
fpdf
requests
pandas
flask
gunicorn
python-dotenv
apscheduler
paynow
pycryptodome
